Able to generate basic alpine Dockerfile
