+++ hacks/test-commands.sh
@@ -1,23 +1,25 @@
 #!/bin/bash
 set -eo pipefail
 
-allbctl --help
-allbctl
-allbctl generate
-allbctl generate ansible
-allbctl generate ansible init
-allbctl generate ansible config
-allbctl generate ansible inventory
-allbctl generate ansible hostVar
-allbctl generate ansible groupVar
-allbctl generate ansible role
-allbctl generate dockerfile ansible
-allbctl generate git
-allbctl generate golang
-allbctl generate java
-allbctl generate kubernetes
-allbctl generate node
-allbctl generate python
-allbctl generate ruby
-allbctl generate scala
-allbctl generate shell
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+
+"${SCRIPT_DIR}"/../bin/allbctl --help
+"${SCRIPT_DIR}"/../bin/allbctl
+"${SCRIPT_DIR}"/../bin/allbctl generate
+"${SCRIPT_DIR}"/../bin/allbctl generate ansible
+"${SCRIPT_DIR}"/../bin/allbctl generate ansible init
+"${SCRIPT_DIR}"/../bin/allbctl generate ansible config
+"${SCRIPT_DIR}"/../bin/allbctl generate ansible inventory
+"${SCRIPT_DIR}"/../bin/allbctl generate ansible hostVar
+"${SCRIPT_DIR}"/../bin/allbctl generate ansible groupVar
+"${SCRIPT_DIR}"/../bin/allbctl generate ansible role
+"${SCRIPT_DIR}"/../bin/allbctl generate dockerfile ansible
+"${SCRIPT_DIR}"/../bin/allbctl generate git
+"${SCRIPT_DIR}"/../bin/allbctl generate golang
+"${SCRIPT_DIR}"/../bin/allbctl generate java
+"${SCRIPT_DIR}"/../bin/allbctl generate kubernetes
+"${SCRIPT_DIR}"/../bin/allbctl generate node
+"${SCRIPT_DIR}"/../bin/allbctl generate python
+"${SCRIPT_DIR}"/../bin/allbctl generate ruby
+"${SCRIPT_DIR}"/../bin/allbctl generate scala
+"${SCRIPT_DIR}"/../bin/allbctl generate shell
